moved RenderHTMLStage from basic to web
