+++ stages/processing/web/pom.xml
@@ -35,6 +35,16 @@
 			<version>1.9.0-rc1</version>
 			<scope>test</scope>
 		</dependency>
+        <dependency>
+            <groupId>net.htmlparser.jericho</groupId>
+            <artifactId>jericho-html</artifactId>
+            <version>3.3</version>
+        </dependency>
+        <dependency>
+            <groupId>commons-io</groupId>
+            <artifactId>commons-io</artifactId>
+            <version>2.4</version>
+        </dependency>		
 	</dependencies>
 
 	<build>
